+++ api/src/main/java/cc/carm/lib/easysql/api/builder/DeleteBuilder.java
@@ -2,7 +2,7 @@ package cc.carm.lib.easysql.api.builder;
 
 import cc.carm.lib.easysql.api.SQLAction;
 
-public interface DeleteBuilder extends ConditionalBuilder<DeleteBuilder, SQLAction<Integer>> {
+public interface DeleteBuilder extends ConditionalBuilder<DeleteBuilder, SQLAction<Long>> {
 
     String getTableName();
